--- src/components/home/IntroductionVideo.tsx
@@ -1,102 +1,0 @@
-import React, {useState, useEffect, useRef, SVGProps} from "react"
-
-const VideoBackground = (props: SVGProps<SVGSVGElement>) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    xmlnsXlink="http://www.w3.org/1999/xlink"
-    fill="none"
-    viewBox="0 0 1728 972"
-    {...props}
-  >
-    <g clipPath="url(#a123)">
-      <path fill="#1F1E16" d="M0 0h1728v972H0z" />
-      <path fill="url(#b123)" d="M-793.8 0H1728v972H-793.8z" />
-      <path
-        fill="#FFE82F"
-        fillOpacity={0.4}
-        d="m697 1061.2 327.7-712.9a64.8 64.8 0 0 0-14.8-74.4L563.5-142a64.8 64.8 0 0 0-47.5-17.2l-622 31.8a64.8 64.8 0 0 0-61.4 64.8V1034c0 35.8 29 64.8 64.8 64.8h740.7a64.8 64.8 0 0 0 58.9-37.7Z"
-      />
-      <path
-        fill="#FFE82F"
-        d="m633.5 1082.8 327.7-712.9a64.8 64.8 0 0 0-14.7-74.5L500-120.4a64.8 64.8 0 0 0-47.5-17.2l-621.9 31.8A64.8 64.8 0 0 0-230.8-41v1096.7c0 35.8 29 64.8 64.8 64.8h740.6a64.8 64.8 0 0 0 58.9-37.7"
-      />
-      <path
-        fill="#000"
-        d="M147.7 335.9v-15h12v-60.7h-12v-15h46.7c5.5 0 10.3 1 14.4 2.8a21.3 21.3 0 0 1 13 20.1v1.3c0 4.2-.7 7.6-2.3 10.3a19 19 0 0 1-5.6 6 24 24 0 0 1-6 3v2.4c2 .5 4 1.5 6.3 3a17 17 0 0 1 5.7 6c1.6 2.7 2.5 6.2 2.5 10.5v1.3c0 5-1.2 9.4-3.5 13a22.6 22.6 0 0 1-9.8 8.1c-4 2-8.8 2.9-14.2 2.9zm29-15.6h16.1q5.55 0 9-2.7a9.3 9.3 0 0 0 3.5-7.7v-1.3c0-3.4-1.2-6-3.4-7.8a14 14 0 0 0-9-2.7h-16.2zm0-37.8h15.9c3.5 0 6.4-.9 8.6-2.7a9 9 0 0 0 3.5-7.5V271a9 9 0 0 0-3.3-7.5 13.6 13.6 0 0 0-8.8-2.8h-15.9zm79.5 54.4c-5 0-9.4-1-13.2-3.3a24 24 0 0 1-8.7-9.6c-2-4-3.1-8.7-3.1-14v-38.4h16.3v37c0 5 1.2 8.6 3.5 11q3.6 3.6 10.2 3.6c5 0 9-1.6 11.7-5 2.8-3.3 4.2-8 4.2-14v-32.6h16.3V336h-16v-8.4H275c-1 2.1-3 4.3-5.8 6.3a22 22 0 0 1-13 3.1Zm50.2-1v-64.3h16.3V336h-16.3Zm8.1-71.8c-2.9 0-5.4-1-7.5-2.8a10 10 0 0 1-3-7.6 10 10 0 0 1 10.5-10.3c3 0 5.6 1 7.6 2.8 2 2 3 4.4 3 7.5a10 10 0 0 1-10.6 10.4m21.1 71.8v-90.7H352v90.7h-16.3Zm55.2 1.8a29 29 0 0 1-14.4-3.7c-4.4-2.6-8-6.4-10.6-11.3a37 37 0 0 1-4-17.9v-2c0-7 1.3-13 4-18a27.9 27.9 0 0 1 25-15c3.9 0 7.1.5 9.7 1.4 2.7.9 4.8 2 6.5 3.4a16 16 0 0 1 3.7 4.4h2.4v-33.8h16.3v90.7h-16v-7.8H411a21 21 0 0 1-6.9 6.6c-3 2-7.4 3-13.3 3m4.9-14.2q7.5 0 12.6-4.8c3.3-3.3 5-8 5-14.3v-1.3c0-6.2-1.7-11-5-14.1a17 17 0 0 0-12.6-5c-5 0-9.2 1.7-12.6 5-3.3 3.2-5 7.9-5 14.1v1.3c0 6.2 1.7 11 5 14.3q5.1 4.8 12.6 4.8m-214 144.2a31.8 31.8 0 0 1-29-16.5q-4.5-8.1-4.5-19.8v-21.7c0-11.7 3.3-20.7 9.9-26.9a37 37 0 0 1 26.7-9.4q16.65 0 25.8 9a31.6 31.6 0 0 1 9.2 23.9v.5h-16.9v-1c0-3.2-.7-6.1-2-8.7a14 14 0 0 0-6-6.1c-2.7-1.6-6-2.3-10.1-2.3-6 0-10.8 1.8-14.3 5.5a21.4 21.4 0 0 0-5.2 15.2v22.3c0 6.3 1.8 11.4 5.2 15.3 3.5 3.8 8.3 5.7 14.5 5.7 6.3 0 10.8-1.7 13.6-5 2.9-3.2 4.3-7.4 4.3-12.4V434h-21.5v-14.5h38.4v46.4H204v-8.7h-2.4c-.6 1.5-1.6 3-3 4.7-1.3 1.6-3.2 3-6 4.1a29.2 29.2 0 0 1-11 1.7Zm49.7-1.8v-64.3h16.1v7.3h2.4c.9-2.6 2.5-4.5 4.6-5.7q3.45-1.8 7.8-1.8h7.8v14.5h-8c-4.2 0-7.6 1.1-10.3 3.3s-4 5.6-4 10.2v36.5zm64.7 1.8a31 31 0 0 1-12.3-2.3c-3.7-1.7-6.5-4-8.7-7-2-3.1-3.1-6.9-3.1-11.3s1-8 3-11c2.3-3 5.2-5.3 8.9-6.7a34 34 0 0 1 13-2.4h17.6v-3.6c0-3-1-5.5-2.9-7.4q-2.85-3-9-3c-4 0-7.1 1-9.1 2.9-2 1.8-3.3 4.2-3.9 7.1l-15-5a24.5 24.5 0 0 1 14.1-15.6c3.9-1.7 8.6-2.6 14.1-2.6 8.5 0 15.2 2.1 20.1 6.4 5 4.2 7.4 10.3 7.4 18.4v24c0 2.5 1.2 3.8 3.6 3.8h5.2V466h-10.9c-3.2 0-5.8-.8-7.9-2.3a7.4 7.4 0 0 1-3-6.3h-2.6a15.6 15.6 0 0 1-8 8.6 24 24 0 0 1-10.6 1.8Zm2.8-13.2c4.6 0 8.3-1.3 11.2-3.8 3-2.6 4.4-6 4.4-10.2v-1.3H298c-3 0-5.4.7-7.1 2a6.4 6.4 0 0 0-2.6 5.4 7 7 0 0 0 2.7 5.7c1.8 1.5 4.5 2.2 8 2.2Zm47 37.3v-90.2h16v7.8h2.4c1.4-2.5 3.7-4.7 6.8-6.6 3.2-2 7.6-3 13.4-3a27.9 27.9 0 0 1 25 15 37 37 0 0 1 4 18v2c0 7-1.3 13-4 17.9-2.7 5-6.2 8.7-10.6 11.3a28.6 28.6 0 0 1-14.4 3.7c-3.9 0-7.2-.5-9.9-1.4-2.6-.9-4.7-2-6.3-3.4s-2.8-3-3.8-4.4h-2.3v33.3zm33.7-38.3c5 0 9.3-1.6 12.5-4.8 3.4-3.3 5-8 5-14.3v-1.3c0-6.2-1.6-11-5-14.1-3.3-3.3-7.5-5-12.5-5s-9.2 1.7-12.6 5c-3.4 3.2-5 7.9-5 14.1v1.3c0 6.2 1.6 11 5 14.3q5.1 4.8 12.6 4.8m43.7 12.4v-90.7h16.4v34.3h2.3c.7-1.4 1.8-2.7 3.2-4.1 1.5-1.4 3.4-2.5 5.9-3.4a27 27 0 0 1 9.4-1.4 24 24 0 0 1 22 13c2 4 3 8.6 3 14v38.3h-16.3v-37c0-4.9-1.2-8.5-3.6-11-2.4-2.4-5.7-3.6-10.1-3.6-5 0-9 1.7-11.7 5a21 21 0 0 0-4.1 14v32.6zm109.4 1.8a38.6 38.6 0 0 1-27.2-9.3c-6.7-6.3-10-15.3-10-27v-21.7c0-11.7 3.3-20.7 10-26.9a38.2 38.2 0 0 1 27.2-9.4c11.4 0 20.5 3.1 27.3 9.4 6.7 6.2 10 15.2 10 26.9v21.7c0 11.7-3.3 20.7-10 27a38.6 38.6 0 0 1-27.3 9.3m0-15.3c6.4 0 11.4-1.8 15-5.6 3.5-3.7 5.3-8.6 5.3-14.9v-22.8c0-6.2-1.8-11.2-5.4-14.9a19.6 19.6 0 0 0-14.9-5.5c-6.3 0-11.2 1.8-14.9 5.5a20.7 20.7 0 0 0-5.3 15v22.7c0 6.3 1.8 11.2 5.3 15 3.7 3.7 8.6 5.5 15 5.5Zm6 36.8a14 14 0 0 1-10.4-4c-2.6-2.6-3.9-6-3.9-10.5V466H541v6.2c0 2.6 1.2 4 3.7 4h7.6v13.1h-13.6Zm41.2-23.3v-90.7h17v75.1h41.6V466h-58.6ZM144 595.9l23.8-90.7h29.8l23.9 90.7H204l-5-20h-32.4l-4.9 20h-17.6Zm26.6-35.8H195L184 516h-2.3l-11 44.2Zm56.6 35.8v-90.7h37.3c5.7 0 10.7 1.1 15 3.5 4.4 2.2 7.8 5.4 10.3 9.6a28 28 0 0 1 3.7 14.7v1.9q0 8.4-3.9 14.7c-2.5 4.2-6 7.4-10.3 9.8a31.5 31.5 0 0 1-14.8 3.3h-20.2V596h-17.1Zm17-48.7H263c4 0 7.3-1.2 9.8-3.4 2.6-2.2 3.8-5.3 3.8-9.2v-1.3c0-3.9-1.2-7-3.8-9.2a14.2 14.2 0 0 0-9.8-3.4h-18.5v26.5Zm56 48.7v-90.7h17.1v90.7zm56.2 1.8a36 36 0 0 1-20.6-5.4 22.4 22.4 0 0 1-9.7-15.6l15-3.9q.9 4.5 3 7.2c1.5 1.7 3.2 3 5.3 3.7 2.2.7 4.5 1 7 1 3.8 0 6.6-.6 8.4-1.9 1.8-1.4 2.7-3 2.7-5a5 5 0 0 0-2.6-4.6c-1.7-1.1-4.4-2-8.2-2.7l-3.7-.6c-4.5-1-8.6-2-12.3-3.5a23 23 0 0 1-9-6.4 15.6 15.6 0 0 1-3.3-10.4 17 17 0 0 1 7-14.6 30.2 30.2 0 0 1 18.4-5.2c7.2 0 13.1 1.6 17.9 4.8a22 22 0 0 1 9.3 12.6l-15.1 4.6c-.7-3.2-2.2-5.6-4.3-7-2-1.3-4.7-2-7.8-2s-5.5.5-7.1 1.6a5 5 0 0 0-2.5 4.5 5 5 0 0 0 2.6 4.6c1.7 1 4 1.7 7 2.2l3.6.7c4.9.8 9.2 2 13.1 3.5 4 1.4 7.1 3.4 9.4 6 2.3 2.7 3.5 6.2 3.5 10.7 0 6.7-2.5 12-7.4 15.7a31.9 31.9 0 0 1-19.6 5.4M138.9 751.8v-14h8.6q3.6 0 3.6-3.9v-72.3h16.3v75.7c0 4.4-1.3 8-3.9 10.5-2.6 2.7-6 4-10.3 4zm20.3-97.7c-2.9 0-5.4-1-7.5-2.8a10 10 0 0 1-3-7.6 10 10 0 0 1 10.5-10.3c3 0 5.6 1 7.6 2.8 2 2 3 4.4 3 7.5a10 10 0 0 1-10.6 10.4m51.7 73.6c-6.4 0-12.1-1.3-17.2-3.9a29.2 29.2 0 0 1-12-11.2 34 34 0 0 1-4.5-17.8v-2a34 34 0 0 1 4.4-17.8c3-5 7-8.7 12-11.3a37.5 37.5 0 0 1 17.3-3.9c6.4 0 12.1 1.3 17.2 3.9s9.2 6.3 12 11.3a34 34 0 0 1 4.5 17.7v2.1a34 34 0 0 1-4.4 17.8c-3 4.9-7 8.6-12 11.2a37.5 37.5 0 0 1-17.3 4Zm0-14.5c5 0 9.2-1.6 12.4-4.8 3.3-3.3 5-8 5-14v-1.3c0-6-1.7-10.7-5-13.9-3.2-3.2-7.3-4.9-12.4-4.9-5 0-9.2 1.7-12.4 5-3.3 3.1-5 7.8-5 13.8v1.3c0 6 1.7 10.7 5 14 3.2 3.2 7.4 4.8 12.4 4.8m50.6 38.6v-14.2h35c2.4 0 3.6-1.3 3.6-4v-16.1h-2.3a16 16 0 0 1-9.2 8c-2.5 1-5.7 1.4-9.6 1.4-5 0-9.4-1-13.2-3.3a24 24 0 0 1-8.7-9.6c-2.1-4-3.1-8.7-3.1-14v-38.4h16.3v37c0 5 1.2 8.6 3.5 11q3.6 3.6 10.2 3.6c5 0 9-1.6 11.7-5 2.8-3.3 4.1-8 4.1-14v-32.6h16.4v75.7c0 4.4-1.3 8-4 10.5-2.5 2.7-6 4-10.3 4zm78.3-25.9v-50.8h-16.3v-13.5h16.3v-11.9c0-4.4 1.3-8 3.9-10.5a14 14 0 0 1 10.3-4h17v13.5h-11.2c-2.4 0-3.6 1.3-3.6 3.8v9.1H373v13.5h-17v50.8h-16.3Zm65.1 1c-5 0-9.4-1-13.2-3.3a24 24 0 0 1-8.7-9.6c-2-4-3-8.7-3-14v-38.4h16.2v37c0 5 1.2 8.6 3.5 11 2.5 2.4 5.9 3.6 10.3 3.6 5 0 8.9-1.6 11.6-5 2.8-3.3 4.2-8 4.2-14v-32.6H442V726h-16v-8.4h-2.4c-1 2.1-3 4.3-5.8 6.3a22 22 0 0 1-13 3.1Zm50.2-1v-90.7h16.3v90.7zm29.2 0v-90.7h16.4v90.7zm36.3 26v-14.3h35c2.4 0 3.6-1.3 3.6-4v-16.1H557a16 16 0 0 1-9.2 8c-2.5 1-5.7 1.4-9.6 1.4-5 0-9.4-1-13.2-3.3a24 24 0 0 1-8.7-9.6c-2.1-4-3.1-8.7-3.1-14v-38.4h16.3v37c0 5 1.2 8.6 3.5 11q3.6 3.6 10.2 3.6c5 0 9-1.6 11.7-5 2.8-3.3 4.1-8 4.1-14v-32.6h16.4v75.7c0 4.4-1.3 8-4 10.5-2.5 2.7-6 4-10.3 4h-40.4Z"
-      />
-      <path
-        fill="#121315"
-        fillRule="evenodd"
-        d="M197.1 112.5h-36.6c-4.5 0-7.6 0-9.9.3-2.2.3-3.2.8-3.9 1.5s-1.2 1.7-1.5 4c-.3 2.2-.3 5.3-.3 9.8v9.3c0 4.5 0 7.6.3 9.9.3 2.2.8 3.2 1.5 3.9s1.7 1.2 4 1.5c2.2.3 5.3.3 9.8.3H197c4.5 0 7.6 0 9.9-.3 2.2-.3 3.2-.8 3.9-1.5s1.2-1.7 1.5-4c.3-2.2.3-5.3.3-9.8v-9.3c0-4.5 0-7.6-.3-9.9-.3-2.2-.8-3.2-1.5-3.9s-1.7-1.2-4-1.5c-2.2-.3-5.3-.3-9.8-.3Zm-52.6-.4c-2.8 2.7-2.8 7.1-2.8 16v9.3c0 8.9 0 13.3 2.8 16 2.7 2.8 7.1 2.8 16 2.8H197c8.9 0 13.3 0 16-2.8 2.8-2.7 2.8-7.1 2.8-16v-9.3c0-8.9 0-13.3-2.8-16-2.7-2.7-7.1-2.7-16-2.7h-36.6c-8.9 0-13.3 0-16 2.7Z"
-        clipRule="evenodd"
-      />
-      <path
-        fill="#121315"
-        fillRule="evenodd"
-        d="M160.7 121.7c.7-.7 2-.7 2.6 0l9.8 9.7c.7.8.7 2 0 2.7l-9.8 9.7a1.9 1.9 0 0 1-2.6-2.6l8.4-8.4-8.4-8.5c-.7-.7-.7-1.9 0-2.6"
-        clipRule="evenodd"
-      />
-      <path
-        fill="#121315"
-        d="M178.5 142.6c0-1 .8-2 1.9-2h16a1.9 1.9 0 0 1 0 3.9h-16c-1 0-1.9-1-1.9-2Zm51.3 1.6c-1.1 0-2-.3-2.7-1-.6-.6-1-1.5-1-2.7v-9.1h-4V128h4v-5h4.2v5h4.4v3.4h-4.4v8.4c0 .7.3 1 .9 1h3.1v3.4zm12.4.5c-1.2 0-2.2-.2-3.1-.6-1-.4-1.7-1-2.2-1.8a5 5 0 0 1-.8-2.8c0-1.1.3-2 .8-2.8s1.3-1.3 2.2-1.7c1-.4 2-.6 3.3-.6h4.5v-1c0-.7-.3-1.3-.8-1.8-.4-.5-1.2-.7-2.3-.7-1 0-1.7.2-2.2.7-.6.4-.9 1-1 1.8l-3.8-1.3c.2-.8.6-1.6 1.2-2.3s1.4-1.2 2.3-1.6c1-.5 2.2-.7 3.6-.7 2.2 0 3.8.5 5 1.6 1.3 1 2 2.6 2 4.7v6c0 .7.3 1 .9 1h1.3v3.4h-2.8c-.8 0-1.4-.2-2-.6-.5-.4-.8-.9-.8-1.5h-.6c0 .2-.3.5-.6 1a4 4 0 0 1-1.4 1 6 6 0 0 1-2.7.6m.7-3.4c1.2 0 2.1-.3 2.8-1 .8-.6 1.2-1.4 1.2-2.5v-.3h-4.2a3 3 0 0 0-1.8.5c-.4.3-.7.7-.7 1.3s.3 1.1.7 1.5c.5.3 1.2.5 2 .5m11.9 2.9V128h4.1v16.2zm2-18.1c-.7 0-1.3-.3-1.8-.8-.5-.4-.8-1-.8-1.9a2.5 2.5 0 0 1 2.7-2.6c.7 0 1.4.3 1.9.7.5.5.7 1.2.7 2 0 .7-.2 1.4-.7 1.8-.5.5-1.2.8-2 .8m5.4 18.1v-23h4.1v23zm15 .5c-1.6 0-3-.4-4.3-1a7.4 7.4 0 0 1-3-2.9 8.7 8.7 0 0 1-1.1-4.5v-.4c0-1.8.4-3.3 1.1-4.6a7.4 7.4 0 0 1 3-2.8c1.3-.7 2.7-1 4.3-1a9 9 0 0 1 4 .8 7.6 7.6 0 0 1 4 5.5l-4 .9c0-.7-.2-1.3-.5-1.8s-.8-1-1.3-1.3c-.6-.3-1.3-.4-2-.4-1 0-1.7.2-2.4.5-.6.4-1.1 1-1.5 1.6-.4.7-.6 1.6-.6 2.6v.4c0 1 .2 1.9.6 2.6s.9 1.2 1.5 1.6c.7.3 1.5.5 2.3.5 1.3 0 2.2-.3 2.8-1 .7-.6 1-1.5 1.3-2.5l4 1a9 9 0 0 1-1.5 3.1 7 7 0 0 1-2.7 2.3 9 9 0 0 1-4 .8m15.3 0c-1.2 0-2.2-.2-3.1-.6-1-.4-1.7-1-2.2-1.8a5 5 0 0 1-.8-2.8c0-1.1.3-2 .8-2.8s1.3-1.3 2.2-1.7c1-.4 2-.6 3.3-.6h4.5v-1c0-.7-.3-1.3-.8-1.8-.4-.5-1.2-.7-2.3-.7-1 0-1.8.2-2.3.7-.5.4-.8 1-1 1.8l-3.7-1.3c.2-.8.6-1.6 1.2-2.3s1.4-1.2 2.3-1.6c1-.5 2.2-.7 3.6-.7 2.1 0 3.8.5 5 1.6 1.3 1 2 2.6 2 4.7v6c0 .7.3 1 .9 1h1.3v3.4h-2.8c-.8 0-1.4-.2-2-.6-.5-.4-.8-.9-.8-1.5h-.6c0 .2-.3.5-.6 1a4 4 0 0 1-1.4 1 6 6 0 0 1-2.7.6m.7-3.4c1.2 0 2.1-.3 2.8-1 .8-.6 1.2-1.4 1.2-2.5v-.3H293a3 3 0 0 0-1.8.5c-.4.3-.7.7-.7 1.3s.3 1.1.7 1.5c.5.3 1.2.5 2 .5m11.8 2.9v-23h4.2v23h-4.1Zm7.5 0v-23h4.1v23z"
-      />
-    </g>
-    <defs>
-      <clipPath id="a123">
-        <path fill="#fff" d="M0 0h1728v972H0z" />
-      </clipPath>
-      <pattern id="b123" width={1} height={1} patternContentUnits="objectBoundingBox">
-        <use xlinkHref="#c" transform="scale(0)" />
-      </pattern>
-    </defs>
-  </svg>
-)
-
-const IntroductionVideo: React.FC = () => {
-  const [isPlaying, setIsPlaying] = useState(false)
-  const [thumbnailUrl, setThumbnailUrl] = useState("")
-  const videoId = "1011521201"
-  const videoRef = useRef<HTMLDivElement>(null)
-
-  useEffect(() => {
-    const observer = new IntersectionObserver(
-      ([entry]) => {
-        if (entry.isIntersecting) {
-          setIsPlaying(true)
-        }
-      },
-      {rootMargin: "200px"},
-    )
-
-    if (videoRef.current) {
-      observer.observe(videoRef.current)
-    }
-
-    return () => {
-      observer.disconnect()
-    }
-  }, [])
-
-  return (
-    <div className="video-wrapper" ref={videoRef}>
-      <div className="video-container">
-        <VideoBackground className="video-background" />
-        {!isPlaying ? (
-          <div className="play-button" onClick={() => setIsPlaying(true)}>
-            <img className="size-6 md:size-10" src="/icons/basic/play.svg" alt="Play Button Icon" />
-          </div>
-        ) : (
-          <iframe
-            src={`https://player.vimeo.com/video/${videoId}?autoplay=0&badge=0&autopause=0&player_id=0&app_id=58479`}
-            allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture"
-            allowFullScreen
-            className="absolute top-0 left-0 w-full h-full"
-            title="Tailcall Introduction Video"
-          />
-        )}
-      </div>
-    </div>
-  )
-}
-
-export default IntroductionVideo
