+++ src/components/home/IntroductionVideo/style.css
@@ -1,0 +1,70 @@
+.video-wrapper {
+  padding: 1.75rem;
+  background-color: #fff;
+  position: relative;
+  overflow: hidden;
+}
+
+.video-wrapper::before {
+  content: "";
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background-image: linear-gradient(90deg, #c5c5c5 1px, transparent 1px), linear-gradient(#c5c5c5 1px, transparent 1px);
+  background-size: 100px 50px;
+  background-position: -1px -1px;
+  opacity: 0.5;
+  max-width: 1595px;
+  margin-left: auto;
+  margin-right: auto;
+}
+
+.video-container {
+  position: relative;
+  width: 100%;
+  overflow: hidden;
+  padding-bottom: 56.25%;
+  border-radius: 0.375rem;
+  z-index: 10;
+  box-shadow: 0px 4px 32px 0px #ffe933;
+}
+
+.video-background {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background-size: cover;
+  background-position: center;
+}
+
+@media (min-width: 768px) {
+  .video-wrapper {
+    padding: 7rem /* 112px */;
+  }
+  .video-container {
+    border-radius: 1.5rem /* 24px */;
+  }
+}
+
+@media (min-width: 1536px) {
+  .video-container {
+    max-width: 80rem;
+    margin-left: auto;
+    margin-right: auto;
+    aspect-ratio: 16/9;
+    padding-bottom: 0;
+  }
+  .video-wrapper::after {
+    content: "";
+    position: absolute;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    background: radial-gradient(ellipse at center, transparent 50%, white 85%);
+  }
+}
