+++ src/css/custom.css
@@ -698,53 +698,3 @@ span.token.tag.script.language-javascript {
   height: 100%;
   border: 0;
 }
-
-.video-wrapper {
-  @apply p-7 md:p-28 bg-white relative overflow-hidden;
-}
-
-.video-wrapper::before {
-  content: "";
-  position: absolute;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background-image: linear-gradient(90deg, #c5c5c5 1px, transparent 1px), linear-gradient(#c5c5c5 1px, transparent 1px);
-  background-size: 100px 50px;
-  background-position: -1px -1px;
-  opacity: 0.5;
-  max-width: 1595px;
-  margin-left: auto;
-  margin-right: auto;
-}
-
-.video-container {
-  @apply relative w-full overflow-hidden pb-[56.25%] rounded-md md:rounded-3xl z-10;
-  box-shadow: 0px 4px 32px 0px #ffe933;
-}
-
-.video-background {
-  @apply absolute top-0 left-0 w-full h-full bg-cover bg-center;
-}
-
-.play-button {
-  @apply absolute inset-0 flex items-center justify-center cursor-pointer absolute top-1/2 left-1/2 transform -translate-x-1/2 -translate-y-1/2;
-  @apply w-14 h-14 md:w-[120px] md:h-[120px] bg-white border md:border-[3px] border-solid border-tailCall-yellow;
-  @apply rounded-[20px] md:rounded-[42px] shadow-md backdrop-blur-xl;
-}
-
-@media (min-width: 1536px) {
-  .video-container {
-    @apply max-w-7xl mx-auto aspect-video pb-0;
-  }
-  .video-wrapper::after {
-    content: "";
-    position: absolute;
-    top: 0;
-    left: 0;
-    right: 0;
-    bottom: 0;
-    background: radial-gradient(ellipse at center, transparent 50%, white 85%);
-  }
-}
